Update fines in task description
Add more specific instructions for rating for task. Because students will not be able to access the application code, it makes sense to write a separate paragraph on the basis of this instruction - an instruction for checking the student’s task, cutting out all the nuances about the code from there.