doc: update first criteria of number 3 (criteria to CSS) to be more clear
unfortunately, a lot of reviewers consider the phrase "при уменьшении масштаба" to make the width of the page less than 1280px, and therefore they put less point as on the width of the screen less than 1280 there are some problems (as there is no adaptive design yet). Please make it more clear for reviewers. Thank you!